% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert}
\alias{insert}
\title{Insert Function}
\usage{
insert(
  conn = NULL,
  insert_table = NULL,
  insert_vector = NULL,
  insert_values = NULL
)
}
\arguments{
\item{conn}{connection to SQLite database}

\item{insert_table}{table name into which values are to be inserted}

\item{insert_vector}{column names(vector) into which values should be inserted}

\item{insert_values}{values(vector) which are to be inserted}
}
\description{
The function inserts values into a table.
}
\examples{
insert()
}
\keyword{insert}
