% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_rsqliteadmin}
\alias{run_rsqliteadmin}
\title{Run the Shiny Application}
\usage{
run_rsqliteadmin(...)
}
\arguments{
\item{...}{A series of options to be used inside the app.}
}
\description{
Run the Shiny Application
}
