% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search}
\alias{search}
\title{Search Function}
\usage{
search(
  conn = NULL,
  distinct = FALSE,
  display_columns = NULL,
  table_names = NULL,
  search_parameters = NULL,
  search_parameters_values = NULL,
  search_parameters_operators = NULL,
  ordering = NULL,
  ordering_parameters = NULL,
  check_null = NULL,
  check_null_parameters = NULL
)
}
\arguments{
\item{conn}{connection to SQLite database}

\item{distinct}{adds "DISTINCT" in the sql query if true}

\item{display_columns}{columns to be selected}

\item{table_names}{tables from which the columns should be selected}

\item{search_parameters}{column names on the basis of which to filter out results}

\item{search_parameters_values}{values for the respective search parameters}

\item{search_parameters_operators}{supports >, <, ==, >=, <=}

\item{ordering}{adds ORDER BY statement to query to sort the results}

\item{ordering_parameters}{columns by which to sort}

\item{check_null_parameters}{columns which should be tested for IS NULL or IS NOT NULL}

\item{check_null=NULL}{adds a NOT NULL or NULL statement to the query}
}
\description{
The function takes multiple search parameters, creates a query and filters out results on the basis of that query.
}
\examples{
search()
}
\keyword{search}
